Remove unused file and move message_bus to primitives/ directory.
