+++ packages/devtools_app/lib/src/primitives/message_bus.dart
