+++ packages/devtools_app/lib/src/shared/globals.dart
@@ -4,8 +4,8 @@
 
 import '../config_specific/ide_theme/ide_theme.dart';
 import '../config_specific/import_export/import_export.dart';
-import '../core/message_bus.dart';
 import '../extension_points/extensions_base.dart';
+import '../primitives/message_bus.dart';
 import '../primitives/storage.dart';
 import 'framework_controller.dart';
 import 'preferences.dart';
