+++ packages/devtools_app/test/message_bus_test.dart
@@ -2,8 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-
-import 'package:devtools_app/src/core/message_bus.dart';
+import 'package:devtools_app/src/primitives/message_bus.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
