+++ tool/bots.sh
@@ -112,9 +112,9 @@ elif [ "$BOT" = "test_ddc" ]; then
     # The flutter tool doesn't support excluding a specific set of targets,
     # so we explicitly provide them.
     if [ "$PLATFORM" = "vm" ]; then
-        flutter test test/*.dart test/{core,fixtures}/
+        flutter test test/*.dart test/fixtures/
     elif [ "$PLATFORM" = "chrome" ]; then
-        flutter test --platform chrome test/*.dart test/{core,fixtures}/
+        flutter test --platform chrome test/*.dart test/fixtures/
     else
         echo "unknown test platform"
         exit 1
@@ -130,9 +130,9 @@ elif [ "$BOT" = "test_dart2js" ]; then
     # The flutter tool doesn't support excluding a specific set of targets,
     # so we explicitly provide them.
     if [ "$PLATFORM" = "vm" ]; then
-        WEBDEV_RELEASE=true flutter test test/*.dart test/{core,fixtures}/
+        WEBDEV_RELEASE=true flutter test test/*.dart test/fixtures/
     elif [ "$PLATFORM" = "chrome" ]; then
-        WEBDEV_RELEASE=true flutter test --platform chrome test/*.dart test/{core,fixtures}/
+        WEBDEV_RELEASE=true flutter test --platform chrome test/*.dart test/fixtures/
     else
         echo "unknown test platform"
         exit 1
